Support private docker registries on windows

+++ manifests/registry.pp
@@ -47,55 +47,98 @@ define docker::registry(
 
   $docker_command = $docker::params::docker_command
 
+  if $::osfamily == 'windows' {
+    $exec_environment = ['PATH=C:/Program Files/Docker/']
+    $exec_timeout = 3000
+    $exec_path = ['c:/Windows/Temp/', 'C:/Program Files/Docker/']
+    $exec_provider = 'powershell'
+    $password_env = '$env:password'
+    $exec_user = undef
+  } else {
+    $exec_environment = ['HOME=/root']
+    $exec_path = ['/bin', '/usr/bin']
+    $exec_timeout = 0
+    $exec_provider = undef
+    $password_env = "\${password}"
+    $exec_user = $local_user
+  }
+
   if $ensure == 'present' {
     if $username != undef and $password != undef and $email != undef and $version != undef and $version =~ /1[.][1-9]0?/ {
-      $auth_cmd = "${docker_command} login -u '${username}' -p \"\${password}\" -e '${email}' ${server}"
+      $auth_cmd = "${docker_command} login -u '${username}' -p '${password_env}' -e '${email}' ${server}"
       $auth_environment = "password=${password}"
     }
     elsif $username != undef and $password != undef {
-      $auth_cmd = "${docker_command} login -u '${username}' -p \"\${password}\" ${server}"
+      $auth_cmd = "${docker_command} login -u '${username}' -p '${password_env}' ${server}"
       $auth_environment = "password=${password}"
     }
     else {
       $auth_cmd = "${docker_command} login ${server}"
-      $auth_environment = undef
+      $auth_environment = ''
     }
   }
   else {
     $auth_cmd = "${docker_command} logout ${server}"
-    $auth_environment = undef
+    $auth_environment = ''
   }
 
+  $docker_auth = "${title}${auth_environment}${auth_cmd}${local_user}"
   if $receipt {
-    # no - with pw_hash
-    $local_user_strip = regsubst($local_user, '-', '', 'G')
-
-    $_pass_hash = $pass_hash ? {
-      Undef   => pw_hash("${title}${auth_environment}${auth_cmd}${local_user}", 'SHA-512', $local_user_strip),
-      default => $pass_hash
-    }
 
     # server may be an URI, which can contain /
     $server_strip = regsubst($server, '/', '_', 'G')
-    $_auth_command = "${auth_cmd} || rm -f \"/root/registry-auth-puppet_receipt_${server_strip}_${local_user}\""
 
-    file { "/root/registry-auth-puppet_receipt_${server_strip}_${local_user}":
-      ensure  => $ensure,
-      content => $_pass_hash,
-      notify  => Exec["${title} auth"],
+    if $::osfamily != 'windows' {
+      # no - with pw_hash
+      $local_user_strip = regsubst($local_user, '-', '', 'G')
+
+      $_pass_hash = $pass_hash ? {
+        Undef   => pw_hash($docker_auth, 'SHA-512', $local_user_strip),
+        default => $pass_hash
+      }
+      $_auth_command = "${auth_cmd} || rm -f \"/root/registry-auth-puppet_receipt_${server_strip}_${local_user}\""
+
+      file { "/root/registry-auth-puppet_receipt_${server_strip}_${local_user}":
+        ensure  => $ensure,
+        content => $_pass_hash,
+        notify  => Exec["${title} auth"],
+      }
+    } else {
+      $_auth_command = $auth_cmd
+      $pw_hash_path = 'C:/Windows/Temp/compute_hash.ps1'
+      $passfile = "C:/Windows/Temp/registry-auth-puppet_receipt_${server_strip}_${local_user}"
+      file{ $pw_hash_path:
+          ensure  => present,
+          force   => true,
+          content => template('docker/windows/compute_hash.ps1.erb'),
+          notify  => Exec['compute-hash']
+      }
+      exec { 'compute-hash':
+          command     => "& ${pw_hash_path}",
+          provider    => $exec_provider,
+          refreshonly => true,
+          logoutput   => true,
+          notify      => Exec["${title} auth"],
+        }
     }
   }
   else {
     $_auth_command = $auth_cmd
   }
 
+  if $auth_environment != '' {
+    $exec_env = concat($exec_environment, $auth_environment, "docker_auth=${docker_auth}")
+  } else {
+    $exec_env = concat($exec_environment, "docker_auth=${docker_auth}")
+  }
+
   exec { "${title} auth":
-    environment => $auth_environment,
+    environment => $exec_env,
     command     => $_auth_command,
-    user        => $local_user,
-    cwd         => '/root',
-    path        => ['/bin', '/usr/bin'],
-    timeout     => 0,
+    user        => $exec_user,
+    path        => $exec_path,
+    timeout     => $exec_timeout,
+    provider    => $exec_provider,
     refreshonly => $receipt,
   }
